% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxBox.R
\name{MaxBox}
\alias{MaxBox}
\title{MaxBox}
\description{
Regional descriptor method based on the Maximum Box approach of
Eckstein et al. (2002), adapted to search for interpretable regional
descriptors (IRDs).

The method starts from a box covering admissible data around `x_interest`
and uses a branch-and-bound search to find a large homogeneous box.
In this package, the original approach is extended so that the returned
box also contains `x_interest`.
}
\references{
Eckstein, J., Hammer, P. L., Liu, Y., Nediak, M., & Simeone, B. (2002).
The maximum box problem and its application to data analysis.
*Computational Optimization and Applications*, 23(3), 285-298.
\url{https://link.springer.com/article/10.1023/A:1020546910706}

Dandl, S., Casalicchio, G., Bischl, B., & Bothmann, L. (2023).
Interpretable Regional Descriptors: Hyperbox-Based Local Explanations.
\url{https://arxiv.org/abs/2305.02780}
}
\section{Super class}{
\code{\link[irdpackage:RegDescMethod]{irdpackage::RegDescMethod}} -> \code{MaxBox}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MaxBox-new}{\code{MaxBox$new()}}
\item \href{#method-MaxBox-clone}{\code{MaxBox$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="find_box"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-find_box'><code>irdpackage::RegDescMethod$find_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="get_parameters"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-get_parameters'><code>irdpackage::RegDescMethod$get_parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="print"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-print'><code>irdpackage::RegDescMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxBox-new"></a>}}
\if{latex}{\out{\hypertarget{method-MaxBox-new}{}}}
\subsection{Method \code{new()}}{
Creates a new `MaxBox` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaxBox$new(
  predictor,
  strategy = "traindata",
  num_sampled_points = 500L,
  efficient = TRUE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor}) \cr
The object (created with `iml::Predictor$new()`) holding the machine
learning model and the data.}

\item{\code{strategy}}{(`character(1)`) \cr
Data source used during the search.
`"traindata"` uses the available training data.
`"sampled"` uses newly sampled points from the largest local box around
`x_interest`.}

\item{\code{num_sampled_points}}{(`numeric(1)`) \cr
Number of sampled points used when `strategy = "sampled"`.}

\item{\code{efficient}}{(`logical(1)`) \cr
Whether to use the more efficient bound computation described in the
MaxBox approach. Default is `TRUE`.}

\item{\code{quiet}}{(`logical(1)`) \cr
Should information about the optimization status be hidden?
Default is `FALSE`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{RegDesc} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxBox-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MaxBox-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaxBox$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
