% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{PostProcessing}
\alias{PostProcessing}
\title{PostProcessing}
\description{
Proposed post-processing method
}
\section{Super class}{
\code{\link[irdpackage:RegDescMethod]{irdpackage::RegDescMethod}} -> \code{PostProcessing}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PostProcessing-new}{\code{PostProcessing$new()}}
\item \href{#method-PostProcessing-clone}{\code{PostProcessing$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="find_box"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-find_box'><code>irdpackage::RegDescMethod$find_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="print"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-print'><code>irdpackage::RegDescMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostProcessing-new"></a>}}
\if{latex}{\out{\hypertarget{method-PostProcessing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostProcessing$new(
  predictor,
  subbox_relsize = 0.05,
  evaluation_n = 1000,
  paste_alpha = 0.05,
  strategy_ties = "preddist",
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subbox_relsize}}{(`numeric(1)`) \cr Number of proposed values
to search over for numeric features. If increased, more upper and lower values are
inspected. Default = 30 meaning that each subbox covers 1/30 = 3.3 % of whole observed
feature range.}

\item{\code{evaluation_n}}{(`numeric(1)`) \cr The number of samples randomly drawn for a specific subbox in order to evaluate this box.}

\item{\code{paste_alpha}}{(NULL | `numeric(1)`)  Stopping criterion. Minimum fraction of box
allowed for adding. See `Details`.}

\item{\code{strategy_ties}}{(`character(1)`) \cr Strategy to use if
multiple subboxes have 0 impurity in order to break ties. Options are
`'random'` to choose randomly, `'preddist'` to choose box with lowest average
distance to prediction of `x_interest`. Default is `'preddist'`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(RegDesc) Hyperbox
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostProcessing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PostProcessing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostProcessing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
