% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDesc.R
\name{RegDesc}
\alias{RegDesc}
\title{Creates a new Regional Descriptor (`RegDesc`) object.
This method should only be called by the `$find_box` methods of \link{RevPrim}.}
\description{
Creates a new Regional Descriptor (`RegDesc`) object.
This method should only be called by the `$find_box` methods of \link{RevPrim}.

Creates a new Regional Descriptor (`RegDesc`) object.
This method should only be called by the `$find_box` methods of \link{RevPrim}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{box}}{(`data.table`)\cr
The regional descriptor for `x_interest`.}

\item{\code{box_single}}{(`data.table`)\cr
The regional descriptors only mutating a single feature for `x_interest`.}

\item{\code{x_interest}}{(`data.table(1)`) \cr
A single row with the observation of interest.}

\item{\code{method}}{(`character(1)`) \cr
Method with which regional descriptors were generated.}

\item{\code{fixed_features}}{(`character(1)`) \cr
Method with which regional descriptors were generated.}

\item{\code{desired_class}}{(`character(1)`) \cr
Desired class. NULL for regression models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegDesc-new}{\code{RegDesc$new()}}
\item \href{#method-RegDesc-print}{\code{RegDesc$print()}}
\item \href{#method-RegDesc-plot_surface}{\code{RegDesc$plot_surface()}}
\item \href{#method-RegDesc-evaluate}{\code{RegDesc$evaluate()}}
\item \href{#method-RegDesc-evaluate_train}{\code{RegDesc$evaluate_train()}}
\item \href{#method-RegDesc-clone}{\code{RegDesc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDesc-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegDesc-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDesc$new(
  box,
  predictor,
  x_interest,
  desired_range,
  fixed_features = NULL,
  desired_class = NULL,
  method = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{box}}{(`ParamSet`)}

\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with `iml::Predictor$new()`) holding the machine learning model and the data.}

\item{\code{x_interest}}{(`data.table(1)` | `data.frame(1)`) \cr
A single row with the observation of interest.}

\item{\code{desired_range}}{(NULL | `numeric(2)`) \cr
The desired predicted outcome. If NULL (default), the current predicted value of `predictor` for `x_interest` is used as desired prediction.
Alternatively, a vector with two numeric values that specify an outcome interval. This outcome interval needs to include the predicted value of `x_interest`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDesc-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegDesc-print}{}}}
\subsection{Method \code{print()}}{
Prints the `RegDesc` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDesc$print(digits = 2L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDesc-plot_surface"></a>}}
\if{latex}{\out{\hypertarget{method-RegDesc-plot_surface}{}}}
\subsection{Method \code{plot_surface()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDesc$plot_surface(feature_names, grid_size = 250L, surface = "prediction")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDesc-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-RegDesc-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDesc$evaluate(n_samples = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDesc-evaluate_train"></a>}}
\if{latex}{\out{\hypertarget{method-RegDesc-evaluate_train}{}}}
\subsection{Method \code{evaluate_train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDesc$evaluate_train()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDesc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegDesc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDesc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
