% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anchor.R
\name{Anchor}
\alias{Anchor}
\title{Anchors}
\description{
Anchors

Anchors
}
\section{Super class}{
\code{\link[irdpackage:RegDescMethod]{irdpackage::RegDescMethod}} -> \code{Anchor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Anchor-new}{\code{Anchor$new()}}
\item \href{#method-Anchor-clone}{\code{Anchor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="find_box"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-find_box'><code>irdpackage::RegDescMethod$find_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="print"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-print'><code>irdpackage::RegDescMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Anchor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Anchor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Anchor$new(
  predictor,
  tau = 1,
  initialize_bins = TRUE,
  bins = NULL,
  quiet = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with `iml::Predictor$new()`) holding the machine
learning model and the data.}

\item{\code{tau}}{(numeric(1)) The desired precision an anchor needs to achieve.
If no candidate achieves at least this precision, the one with the best
precision will be returned.}

\item{\code{initialize_bins}}{(logical(1)) Whether continuous features should be
discretized into bins.}

\item{\code{bins}}{(list) The discretization bins. Requires one list element for
each feature (same ordering as in predictor$data).}

\item{\code{quiet}}{(logical(1)) Suppress messages.
Immutable features defined in `fixed_features` are automatially set to integer().}

\item{\code{...}}{\cr Further hyperparameters of anchors (see `anchors::anchors`).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(RegDesc) Hyperbox
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Anchor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Anchor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Anchor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
