% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maire.R
\name{Maire}
\alias{Maire}
\title{Maire}
\description{
Maire

Maire
}
\section{Super class}{
\code{\link[irdpackage:RegDescMethod]{irdpackage::RegDescMethod}} -> \code{Maire}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Maire-new}{\code{Maire$new()}}
\item \href{#method-Maire-clone}{\code{Maire$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="find_box"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-find_box'><code>irdpackage::RegDescMethod$find_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="print"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-print'><code>irdpackage::RegDescMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Maire-new"></a>}}
\if{latex}{\out{\hypertarget{method-Maire-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maire$new(
  predictor,
  num_of_iterations = 1000L,
  convergence = FALSE,
  strategy = "traindata",
  num_sampled_points = NULL,
  lambda_val_1 = 20,
  lambda_val = 100L,
  threshold = 0.9999999,
  c1 = 0.4,
  c2 = 15,
  c3 = 0.6,
  c4 = 0.5,
  c5 = 0.5,
  cl = 0.02,
  ch = 0.82,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(`iml::Predictor`) \cr
The object (created with `iml::Predictor$new()`) holding the machine
learning model and the data.}

\item{\code{num_of_iterations}}{(`integer(1)`) \cr Number of iterations for ADAM optimizer.
Default is 10000.}

\item{\code{convergence}}{(`logical(1)`) \cr Whether `num_of_iterations` is conducted after
first convergence, i.e., after precision fell below 1 for the first time}

\item{\code{strategy}}{(`character(1)`)\cr
Either `traindata` using training data or `sampled` using newly sampled data in
ICE curve identified box.}

\item{\code{num_sampled_points}}{(`numeric(1)`)\cr Only considered if `strategy = 'sampled'`.
The number of samples randomly drawn at the beginning.}

\item{\code{lambda_val_1}}{(`numeric(1)`) \cr Weight for precision term (in contrast to coverage),
default is 20.}

\item{\code{lambda_val}}{(`numeric(1)`) \cr Weight for constraint that point to explain
must be contained in box, default 100.}

\item{\code{threshold}}{(`numeric(1)`) \cr Lower threshold for precision default 0.9999999.}

\item{\code{c1}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
should be rather low such that indicator approx has only a small value range (default 0.4).}

\item{\code{c2}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
should be rather high such that there is a steep increase in the indicator approx (default 15).}

\item{\code{c3}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
chosen to provide a step at z = 0 (default 0.6).}

\item{\code{c4}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
chosen such that signum term for approximation is close to 1 (default 0.5).}

\item{\code{c5}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
chosen such that signum term for approximation is close to 1 (default 0.6).}

\item{\code{cl}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
should have a value close to 0 (default 0.02).}

\item{\code{ch}}{(`numeric(1)`) \cr Parameter for approximation of indicator function,
should have a value close to 1 (default 0.82)}

\item{\code{quiet}}{(`logical(1)`) Supress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(RegDesc) Hyperbox
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Maire-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Maire-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maire$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
