% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxBox.R
\name{MaxBox}
\alias{MaxBox}
\title{MaxBox}
\description{
Maximum Box approach of Eckstein.
}
\section{Super class}{
\code{\link[irdpackage:RegDescMethod]{irdpackage::RegDescMethod}} -> \code{MaxBox}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MaxBox-new}{\code{MaxBox$new()}}
\item \href{#method-MaxBox-clone}{\code{MaxBox$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="find_box"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-find_box'><code>irdpackage::RegDescMethod$find_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="print"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-print'><code>irdpackage::RegDescMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxBox-new"></a>}}
\if{latex}{\out{\hypertarget{method-MaxBox-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaxBox$new(
  predictor,
  strategy = "traindata",
  num_sampled_points = 500L,
  efficient = TRUE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with `iml::Predictor$new()`) holding the machine
learning model and the data.}

\item{\code{strategy}}{(`character(1)`)\cr
Either `traindata` using training data or `sampled` using newly sampled data in
ICE curve identified box.}

\item{\code{num_sampled_points}}{(`numeric(1)`)\cr Only considered if `strategy = 'sampled'`.
The number of samples randomly drawn at the beginning.}

\item{\code{efficient}}{(`logical(1)`)\cr Whether the efficient computation
proposed by Eckstein et al. should be used. Default TRUE.}

\item{\code{quiet}}{(`logical(1)`)\cr Should information about the optimization status be hidden? Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(RegDesc) Hyperbox
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxBox-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MaxBox-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaxBox$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
