% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prim.R
\name{Prim}
\alias{Prim}
\title{PRIM}
\description{
PRIM

PRIM
}
\section{Super class}{
\code{\link[irdpackage:RegDescMethod]{irdpackage::RegDescMethod}} -> \code{Prim}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Prim-new}{\code{Prim$new()}}
\item \href{#method-Prim-clone}{\code{Prim$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="find_box"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-find_box'><code>irdpackage::RegDescMethod$find_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="irdpackage" data-topic="RegDescMethod" data-id="print"><a href='../../irdpackage/html/RegDescMethod.html#method-RegDescMethod-print'><code>irdpackage::RegDescMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prim-new"></a>}}
\if{latex}{\out{\hypertarget{method-Prim-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prim$new(
  predictor,
  subbox_relsize = 0.05,
  strategy = "traindata",
  num_sampled_points = 500L,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(`iml::Predictor`) \cr
The object (created with `iml::Predictor$new()`) holding the machine
learning model and the data.}

\item{\code{subbox_relsize}}{(`numeric(1)`) \cr Number of proposed values
to search over for numeric features. If increased, more upper and lower values are
inspected. Default = 0.05 meaning that each subbox covers 5 % of the data points in
a box.}

\item{\code{strategy}}{(`character(1)`)\cr
Either `traindata` using training data or `sampled` using newly sampled data in
ICE curve identified box.}

\item{\code{num_sampled_points}}{(`numeric(1)`)\cr Only considered if `strategy = 'sampled'`.
The number of samples randomly drawn at the beginning.}

\item{\code{quiet}}{(`logical(1)`) Supress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(RegDesc) Hyperbox
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Prim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
