% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDescMethod.R
\name{RegDescMethod}
\alias{RegDescMethod}
\title{Base Class for Regional Descriptor Methods}
\description{
Abstract base class for regional descriptors methods.
}
\section{Inheritance}{

Child classes: \link{RevPrim}, \link{MaireOrig}
}

\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{history}}{(`data.table`) \cr
stores for each iteration of the method the chosen variable, boundary
value and evaluation measures.}

\item{\code{calls_fhat}}{(`integer(1)`)
Number of model prediction calls performed during the search procedure.
Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegDescMethod-new}{\code{RegDescMethod$new()}}
\item \href{#method-RegDescMethod-print}{\code{RegDescMethod$print()}}
\item \href{#method-RegDescMethod-find_box}{\code{RegDescMethod$find_box()}}
\item \href{#method-RegDescMethod-clone}{\code{RegDescMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDescMethod-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegDescMethod-new}{}}}
\subsection{Method \code{new()}}{
Creates a new `RegDescMethod` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDescMethod$new(predictor, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with `iml::Predictor$new()`) holding the machine
learning model and the data.}

\item{\code{quiet}}{(`logical(1)`)\cr Should information about the optimization status be hidden? Default is FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDescMethod-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegDescMethod-print}{}}}
\subsection{Method \code{print()}}{
Prints a `RegDescMethod` object.
The method calls a (private) `$print_parameters()` method which should be implemented by the leaf classes.

Runs the hyperrectangle searching algorithm and returns the hyperrectangles interval ranges.
All observations in the hyperrectangle should have a predicted probability in the interval `desired_prob`
(for classification for a `desired_class`).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDescMethod$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDescMethod-find_box"></a>}}
\if{latex}{\out{\hypertarget{method-RegDescMethod-find_box}{}}}
\subsection{Method \code{find_box()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDescMethod$find_box(
  x_interest,
  desired_range = NULL,
  obsdata = NULL,
  fixed_features = NULL,
  desired_class = NULL,
  box_largest = NULL,
  box_init = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_interest}}{(`data.table(1) | data.frame(1)`) \cr A single row with the observation of interest.}

\item{\code{desired_range}}{(`numeric(2)`) \cr
The desired predicted outcome range - a vector with two numeric values
that specify an outcome interval.
For regression the interval operates on the numeric outcome,
while for classification it reflects either a hard label or
the probability for the class with the largest probability among all
possible outcome classes.}

\item{\code{obsdata}}{(`data.table` | `data.frame`) Data set used to find the box. If NULL (default) either
predictor$data or newly sampled data according to the specified `strategy` is used.}

\item{\code{fixed_features}}{(`character()` | `NULL`) \cr
Names of features that are not allowed to be changed. NULL (default) allows all features to be changed.}

\item{\code{desired_class}}{(`character(1)` | `NULL`) \cr The desired class. Ignored if predictor$task = "regression".
If NULL (default) for a classification task then predictor$class is taken.}

\item{\code{box_largest}}{(`ParamSet`  | `NULL`) \cr
Largest initial box. If NULL, largest box is generated.}

\item{\code{box_init}}{(`ParamSet` | `NULL`) \cr Initial box to process. Ignored if method is not `PostProcessing`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegDescMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegDescMethod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegDescMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
