% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_for_explanation}
\alias{transform_for_explanation}
\title{version 1 = MAIRE paper (0.33 vs. 0.66 = one vs. all)
version 2 = my extended version (dummy encoded features)
version 3 = my MAIRE version (dummy encoded features and [0, 1] for numerical ones) --> hybrid of version 1 and version 2}
\usage{
transform_for_explanation(
  data,
  predictor,
  x_interest,
  version = 2,
  frac = 1e-04
)
}
\description{
version 1 = MAIRE paper (0.33 vs. 0.66 = one vs. all)
version 2 = my extended version (dummy encoded features)
version 3 = my MAIRE version (dummy encoded features and [0, 1] for numerical ones) --> hybrid of version 1 and version 2
}
